Add more tests, fix 1 edge case
It was discovered through testing that RefreshTimeRange will throw if it's called before any sessions have been added, or after the last has been removed.  So far the UI appears to do a perfect job of only calling this method when the object is in a state that supports it, this can easily lead to a bug and probably a crash.  Now the method handles the no sessions case gracefully and resets affected properties to their defaults instead.